MM-48640 Move markMostRecentPostInChannelAsUnread out of mattermost-redux
